dom/security/test/https-first/test_fragment.html

<!DOCTYPE HTML>
<html>
<head>
<title>Bug 1706577: Have https-first mode account for fragment navigations</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>

<script class="testbody" type="text/javascript">
"use strict";
/*
 * Description of the test:
 * Have https-first detect a fragment navigation rather than navigating away
 * from the page.
 */

SimpleTest.waitForExplicitFinish();

const REQUEST_URL = "http://example.com/tests/dom/security/test/https-first/file_fragment.html";
const EXPECT_URL = REQUEST_URL.replace("http://", "https://");

let winTest = null;
let checkButtonClicked = false;

async function receiveMessage(event) {
  let data = event.data;
  if (!checkButtonClicked) {
    ok(data.result == EXPECT_URL, "location is correct");
    ok(data.button, "button is clicked");
    ok(data.info == "onload", "Onloading worked");
    checkButtonClicked = true;
    return;
  }

  // Once the button was clicked we know the tast has finished
  ok(data.button, "button is clicked");
  ok(data.result == "#foo", "location (hash) is correct");
  ok(data.info == "scrolled-to-foo","Scrolled successfully without reloading!");
  window.removeEventListener("message",receiveMessage);
  winTest.close();
  SimpleTest.finish();
}

async function runTest() {
  await SpecialPowers.pushPrefEnv({ set: [
    ["dom.security.https_first", true],
  ]});
  winTest = window.open(REQUEST_URL);
}

window.addEventListener("message", receiveMessage);

runTest();

</script>
</body>
</html>
